Random Contraction
Week 4 (kind of) done.
Instead of solving the assigment, I got excited about trying to
implement graphs. Memory management is a real pain in the ass.
Anyway, for the small test the algorithm works fine, for the large it
causes segmentation fault.
Of course this is a extremelly inefficient way to do this and there are
probably bugs and memory leaks all around, but it was fun.